gateway/requirements.txt
# API Gateway Dependencies
flask==3.1.0
flask-cors==5.0.0
gunicorn==23.0.0
anthropic==0.39.0
openai==1.54.3
requests==2.32.3
python-dotenv==1.0.1
redis==5.2.0
prometheus-client==0.21.0
